% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cities.R
\name{get_cities}
\alias{get_cities}
\title{Get names of cities polled for mayor}
\usage{
get_cities()
}
\value{
\code{data.frame} with information from cities in which there are
  polls for mayoral elections in Brazil and the corresponding years for
  which they are available.

  \itemize{
  \item ano: election year to which the pool corresponds to.
  \item cidade: city where the poll was conducted.
  \item unidade_federativa_nome: federative unit where the city is in.
  \item ambito: acronym of the federative unit where the city is in.
  }
}
\description{
\code{get_cities()} extracts the names for all the cities with corresponding
polls for mayor in Poder 360's poll agreggator API
(\url{https://www.poder360.com.br/pesquisas-de-opiniao/api/}). The function
returns a \code{data.frame} where each observation corresponds to a
combination of city and election year for which there's a poll available.
}
\examples{
# returns a data.frame with all combinations of cities and the election
# years for which they have mayoral polls available
df <- get_cities()
}
